# cmd/api/docs/swagger.yaml
definitions:
...
  types.TestSuiteRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  types.TestSuiteResponse:
    properties:
      assignment_id:
        type: integer
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is the API documentation for the Assembly Visual project.
  termsOfService: http://swagger.io/terms/
  title: Assembly Visual API documentation
  version: "1.0"
paths:
...
  /classroom/{class_id}/assignment/{assignment_id}/test-suite:
    get:
      consumes:
      - application/json
      description: Retrieve all test suites by assignment ID
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.TestSuiteResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-suites
    post:
      consumes:
      - application/json
      description: Add a new test suite
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite Data
        in: body
        name: testSuite
        required: true
        schema:
          $ref: '#/definitions/types.TestSuiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-suites
  /classroom/{class_id}/assignment/{assignment_id}/test-suite/{test_suite_id}:
    delete:
      consumes:
      - application/json
      description: Delete a test suite
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-suites
    get:
      consumes:
      - application/json
      description: Retrieve a test suite by its ID
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TestSuiteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-suites
    put:
      consumes:
      - application/json
      description: Update an existing test suite
      parameters:
      - description: Class ID
        in: path
        name: class_id
        required: true
        type: integer
      - description: Assignment ID
        in: path
        name: assignment_id
        required: true
        type: integer
      - description: Test Suite ID
        in: path
        name: test_suite_id
        required: true
        type: integer
      - description: Test Suite Data
        in: body
        name: testSuite
        required: true
        schema:
          $ref: '#/definitions/types.TestSuiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      tags:
      - test-suites
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"